+++ config/installer_mode_config.php
@@ -8,7 +8,7 @@
 	);
 
 	$count_mode_value = array_sum($mode);
-	$database_dump_file = 'prd_cor_zip';
+	$database_dump_file = 'prd_cor_new';
 	$application_name = $application_colour = $application_login_image = $application_sql_name = $application_db_user_name = $application_mode = $application_theme = $application_favicon = '';
 	
 	if($mode['core'] == true){