+++ plugins/System/src/Model/Table/SystemErrorsTable.php
@@ -16,9 +16,10 @@ class SystemErrorsTable extends AppTable
 {
     public function initialize(array $config): void
     {
-
-        parent::initialize($config);
-        $this->belongsTo('CreatedUser', ['className' => 'Security.Users', 'foreignKey' => 'created_user_id']);
+        if (file_exists(CONFIG . 'app_local.php')) { //POCOR-9203
+            parent::initialize($config);
+            $this->belongsTo('CreatedUser', ['className' => 'Security.Users', 'foreignKey' => 'created_user_id']);
+        }
     }
 
     public function implementedEvents(): array
@@ -35,36 +36,38 @@ class SystemErrorsTable extends AppTable
 
     public function insertError(Throwable $ex)
     {
-        $msg = $ex->getMessage();
-        $trace = $ex->getTraceAsString();
-        $file = $ex->getFile();
-        $line = $ex->getLine();
-        $code = $ex->getCode();
-        $serverInfo = json_encode($_SERVER);
+        if (file_exists(CONFIG . 'app_local.php')) { //POCOR-9203
+            $msg = $ex->getMessage();
+            $trace = $ex->getTraceAsString();
+            $file = $ex->getFile();
+            $line = $ex->getLine();
+            $code = $ex->getCode();
+            $serverInfo = json_encode($_SERVER);
 
-        $clientIp = $ip = $_SERVER['REMOTE_ADDR'];
-        if (isset($_SERVER['HTTP_X_FORWARDED_FOR'])) {
-            $clientIp = $_SERVER['HTTP_X_FORWARDED_FOR'];
+            $clientIp = $ip = $_SERVER['REMOTE_ADDR'];
+            if (isset($_SERVER['HTTP_X_FORWARDED_FOR'])) {
+                $clientIp = $_SERVER['HTTP_X_FORWARDED_FOR'];
+            }
+
+            if($code == 0) { //POCOR-8676
+                $code = 404;
+            }
+            $entity = $this->newEntity([
+                'id' => Text::uuid(),
+                'code' => $code,
+                'error_message' => $msg,
+                'request_method' => $_SERVER['REQUEST_METHOD'],
+                'request_url' => $_SERVER['REQUEST_URI'],
+                'referrer_url' => isset($_SERVER['HTTP_REFERER']) ? $_SERVER['HTTP_REFERER'] : '',
+                'client_ip' => $clientIp,
+                'client_browser' => isset($_SERVER['HTTP_USER_AGENT']) ? $_SERVER['HTTP_USER_AGENT'] : '',
+                'triggered_from' => $file . ' (Line: ' . $line . ')',
+                'stack_trace' => $trace,
+                'server_info' => $serverInfo
+            ]);
+
+            $this->save($entity);
         }
-
-        if($code == 0) { //POCOR-8676
-            $code = 404;
-        }
-        $entity = $this->newEntity([
-            'id' => Text::uuid(),
-            'code' => $code,
-            'error_message' => $msg,
-            'request_method' => $_SERVER['REQUEST_METHOD'],
-            'request_url' => $_SERVER['REQUEST_URI'],
-            'referrer_url' => isset($_SERVER['HTTP_REFERER']) ? $_SERVER['HTTP_REFERER'] : '',
-            'client_ip' => $clientIp,
-            'client_browser' => isset($_SERVER['HTTP_USER_AGENT']) ? $_SERVER['HTTP_USER_AGENT'] : '',
-            'triggered_from' => $file . ' (Line: ' . $line . ')',
-            'stack_trace' => $trace,
-            'server_info' => $serverInfo
-        ]);
-
-        $this->save($entity);
     }
 
     public function beforeFind(Event $event, Query $query, ArrayObject $options, $primary)